Store phone RAM as a float so sub-GB phones can be served
TechAPI #257 corrected storage-as-RAM on 2006-2012 phones (original
iPhone 0.125 GB). SmartphoneRead.ram_gb was int, so the static dump
crashed on those records. Store it as float, type the schema int | float,
and emit whole values as ints so the other ~90k pages keep "ram_gb": 8.

+++ app/models/smartphone.py
@@ -29,7 +29,7 @@ class Smartphone(SQLModel, table=True):
     msrp_usd: int | None = None
 
     # Memory
-    ram_gb: int
+    ram_gb: float  # sub-GB on 2006-2012 phones (original iPhone 0.125)
     storage_options_gb: list[int] = Field(default_factory=list, sa_column=Column(JSON))
     variant: dict[str, Any] = Field(default_factory=dict, sa_column=Column(JSON))
